refactor: separate components according to related folder
