+++ netlify.toml
@@ -1,0 +1,4 @@
+[[redirect]]
+  from = "/*"
+  to = "/index.html"
+  status = 200
