+++ public/index.html
@@ -7,12 +7,8 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <!-- <link
-      href="https://fonts.googleapis.com/css2?family=Rambla:wght@400;700&family=Waiting+for+the+Sunrise&family=WindSong:wght@400;500&display=swap"
-      rel="stylesheet"
-    /> -->
     <link
-      href="https://fonts.googleapis.com/css2?family=Jura&family=Rambla:wght@400;700&family=Waiting+for+the+Sunrise&family=WindSong:wght@400;500&display=swap"
+      href="https://fonts.googleapis.com/css2?family=Rambla:wght@400;700&family=Waiting+for+the+Sunrise&family=WindSong:wght@400;500&display=swap"
       rel="stylesheet"
     />
     <meta name="description" content="Honest Diary Websites" />
