+++ src/components/DetailPage/DiaryDetailAudio.js
@@ -2,8 +2,8 @@ import React, { useState, useRef, useEffect } from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import { EFFECT_MODE, SCRIPT_MODE } from "../constants";
-import ErrorModal from "./common/ErrorModal";
+import { EFFECT_MODE, SCRIPT_MODE } from "../../constants";
+import ErrorModal from "../common/ErrorModal";
 import CustomAudioController from "./CustomAudioController";
 
 function DiaryDetailAudio({ mode, diary }) {
