+++ src/components/EntriesPage/DiarySearch.js
@@ -1,13 +1,13 @@
 import { forwardRef, useState } from "react";
 import DatePicker from "react-datepicker";
 import { AiOutlineSearch } from "react-icons/ai";
+import { addMonths, subDays, subMonths, lightFormat } from "date-fns";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import { addMonths, subDays, subMonths, lightFormat } from "date-fns";
 
-import { DATE_FORMAT, SENTIMENT } from "../constants";
-import StyledButton from "./shared/StyledButton";
-import StyledSelect from "./shared/StyledSelect";
+import { DATE_FORMAT, SENTIMENT } from "../../constants";
+import StyledButton from "../shared/StyledButton";
+import StyledSelect from "../shared/StyledSelect";
 import "react-datepicker/dist/react-datepicker.css";
 
 function Search({ setSearchOptions }) {
