+++ src/components/WritePage/Write.js
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import styled from "styled-components";
 
-import { EFFECT_MODE, SCRIPT_MODE } from "../constants";
-import StyledButton from "./shared/StyledButton";
+import { EFFECT_MODE, SCRIPT_MODE } from "../../constants";
+import StyledButton from "../shared/StyledButton";
 import WriteVoice from "./WriteVoice";
 
 function Write() {
