+++ src/components/common/ErrorModal.js
@@ -1,6 +1,6 @@
+import { AiOutlineCloseCircle } from "react-icons/ai";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import { AiOutlineCloseCircle } from "react-icons/ai";
 
 import Modal from "./Modal";
