+++ src/components/common/SuccessModal.js
@@ -1,6 +1,6 @@
+import { AiOutlineCheckCircle } from "react-icons/ai";
 import PropTypes from "prop-types";
 import styled from "styled-components";
-import { AiOutlineCheckCircle } from "react-icons/ai";
 
 import Modal from "./Modal";
